Create Mark_screen_area_return_topleft_bottomright.Ahk
This is a small script with 25 lines or less that returns the top left XY coordinate and the bottom left bottom right XY coordinate of a screen area that allows you to mark the top left and bottom right corners by clicking the mouse on the top left corner and releasing it on the bottom right corner which displays a great out area while you're selecting it and returns the coordinates as standard numeric variables indicating pixels. It's very simple.. You can activate it by pressing a hotkey that you could list in your help file.